x11-misc/xkeyboard-config: check -L as well
Thought had added this already.

Bug: https://bugs.gentoo.org/957712

+++ x11-misc/xkeyboard-config/xkeyboard-config-2.45-r1.ebuild
@@ -74,7 +74,7 @@ src_install() {
 
 pkg_preinst() {
 	# Avoid touching EROOT if not needed, and use -f just-in-case anyway
-	if [[ -d ${EROOT}/usr/share/X11/xkb ]]; then
+	if [[ -d ${EROOT}/usr/share/X11/xkb && ! -L ${EROOT}/usr/share/X11/xkb ]]; then
 		rm -rf "${EROOT}"/usr/share/X11/xkb || die
 	fi
 	mv "${ED}"/usr/share/X11/xkb{.workaround,} || die
